Refine charts and get them working on anroid
